backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic-settings